+++ ydb/tests/datashard/ya.make
@@ -9,7 +9,6 @@ RECURSE(
     partitioning
     parametrized_queries
     select
-    split_merge
     secondary_index
     s3
     vector_index
