+++ ydb/tests/functional/split_merge/by_load/ya.make
@@ -1,25 +1,22 @@
 PY3TEST()
+
 INCLUDE(${ARCADIA_ROOT}/ydb/tests/harness_dep.inc)
 ENV(YDB_CLI_BINARY="ydb/apps/ydb/ydb")
 
 FORK_SUBTESTS()
 SPLIT_FACTOR(45)
 
-IF (SANITIZER_TYPE OR WITH_VALGRIND)
-    SIZE(LARGE)
-    TAG(ya:fat)
-ELSE()
-    SIZE(MEDIUM)
-ENDIF()
+SIZE(LARGE)
+TAG(ya:fat)
 
 TEST_SRCS(
-    test_split_merge.py
+    conftest.py
+    test_split_merge_by_load.py
 )
 
 PEERDIR(
-    ydb/tests/sql/lib
     ydb/tests/library
-    ydb/tests/datashard/lib
+    ydb/tests/library/fixtures
 )
 
 DEPENDS(
@@ -27,7 +24,3 @@ DEPENDS(
 )
 
 END()
-
-RECURSE(
-    by_load
-)
