+++ ydb/tests/library/fixtures/ya.make
@@ -3,6 +3,7 @@ PY3_LIBRARY()
 PY_SRCS(
     __init__.py
     fulltext.py
+    safe_parametrize.py
 )
 
 PEERDIR(
